fix(mobile): Moshi Long<-double crash, bitmap OOM, keterangan auto-label
Root cause: kolom DB double/decimal (transaksi.nominal, angsuran.*,
gaji_pokok, saldo/bunga profile) dikirim backend sebagai JSON float (mis.
621.54037972656). Mobile DTO declare Long -> Moshi JsonDataException
"expected a long but was 621.54..." -> silent crash di HP (Dashboard,
Angsuran, transaksi list). Device-specific karena nilai pecahan muncul
hanya untuk akun/baris tertentu.

Fix DTO Long? -> Double?, mapper Math.round via helper roundL():
- TransactionDto: TransactionResponse.nominal, TransactionDetailResponse.nominal
- InstallmentDto: angsuranPokok, angsuranBunga, SalaryResponse.gajiPokok
- ProfileDto: 17 field saldo/angsuran/bunga/setoran
- ProfileRepositoryImpl: mapper pakai roundL() (util LoanMath.kt)
- TransactionRepositoryImpl, InstallmentRepositoryImpl: mapper round

Bitmap OOM (silent crash HP low-RAM):
- util/Bitmaps.kt: decodeSampled(bytes, maxPx=1024) pakai inSampleSize
  downsample. Slip gaji/bukti resolusi tinggi (3000x4000 -> 48MB ARGB)
  OOM kill. LoanApplicationScreen decode full-size tiap recompose (tanpa
  remember) -> leak cepat. SavingsFormScreen full-size walau di remember.
- LoanApplicationScreen: toBitmap() private dihapus, pakai decodeSampled()
  + remember(slipBytes). SavingsFormScreen pakai decodeSampled().

Keterangan auto-label (identifikasi pengajuan dari mobile di web):
- LoanApplicationViewModel: submit kirim "Pengajuan Pinjaman Jangka
  Panjang/Pendek/Barang dari Esimko Mobile App dd-MM-yyyy HH:mm" (sebelumnya
  hardcode null -> keterangan kosong di web).
- SavingsFormViewModel: "Pengajuan Setoran dari Esimko Mobile App <ts>",
  append catatan user setelah " — " kalau diisi.

Tambahan: SavingsFormScreen/ViewModel, Bitmaps, LoanMath sebelumnya
untracked (terlewat git add -u di commit sebelumnya) -> masuk sekarang.

+++ EsimkoMobile/app/src/main/java/com/esimko/mobile/data/remote/dto/InstallmentDto.kt
@@ -10,13 +10,16 @@ data class InstallmentResponse(
     @Json(name = "bulan") val bulan: String? = null,
     @Json(name = "nama_bulan") val namaBulan: String? = null,
     @Json(name = "angsuran_ke") val angsuranKe: Int? = 0,
-    @Json(name = "angsuran_bunga") val angsuranBunga: Long? = 0,
-    @Json(name = "angsuran_pokok") val angsuranPokok: Long? = 0,
+    // Double, bukan Long: angsuran.angsuran_pokok/angsuran_bunga kolom DB double.
+    // Moshi menolak pecahan sebagai Long. Dibulatkan di mapper.
+    @Json(name = "angsuran_bunga") val angsuranBunga: Double? = 0.0,
+    @Json(name = "angsuran_pokok") val angsuranPokok: Double? = 0.0,
     @Json(name = "status_angsuran") val statusAngsuran: String? = null,
     @Json(name = "color") val color: String? = null
 )
 
 @JsonClass(generateAdapter = true)
 data class SalaryResponse(
-    @Json(name = "gaji_pokok") val gajiPokok: Long? = 0
+    // Double, bukan Long: gaji_pokok.gaji_pokok kolom DB double. Dibulatkan di mapper.
+    @Json(name = "gaji_pokok") val gajiPokok: Double? = 0.0
 )
